+++ openstack/dcs/v1/lifecycle/BatchDelete.go
@@ -1,0 +1,1 @@
+package lifecycle
