+++ openstack/dcs/v1/lifecycle/Get.go
@@ -1,0 +1,9 @@
+package lifecycle
+
+import "github.com/opentelekomcloud/gophertelekomcloud"
+
+// Get a instance with detailed information by id
+func Get(client *golangsdk.ServiceClient, id string) (r GetResult) {
+	_, r.Err = client.Get(getURL(client, id), &r.Body, nil)
+	return
+}
