+++ openstack/dcs/v1/lifecycle/results.go
@@ -1,4 +1,4 @@
-package instances
+package lifecycle
 
 import (
 	"github.com/opentelekomcloud/gophertelekomcloud"
